+++ src/Encamina.Enmarcha.Core/DataAnnotations/GuidNotEmptyAttribute.cs
@@ -25,7 +25,7 @@ public sealed class GuidNotEmptyAttribute : ValidationAttribute
     /// <remarks>
     /// A null value is accepted as valid since the idea of «Not Empty» doesn't necessarily means «Required».
     /// </remarks>
-    protected override ValidationResult IsValid(object value, ValidationContext validationContext)
+    protected override ValidationResult? IsValid(object value, ValidationContext validationContext)
     {
         return IsValid(value)
             ? ValidationResult.Success