+++ src/Encamina.Enmarcha.Core/DataAnnotations/UriAttribute.cs
@@ -34,7 +34,7 @@ public sealed class UriAttribute : DataTypeAttribute
     /// <remarks>
     /// A null value is accepted as valid since the idea of «Not Empty» doesn't necessarily means «Required».
     /// </remarks>
-    protected override ValidationResult IsValid(object value, ValidationContext validationContext)
+    protected override ValidationResult? IsValid(object value, ValidationContext validationContext)
     {
         return IsValid(value)
             ? ValidationResult.Success